requirements.txt
tifffile
bokeh.palettes
zarr
numpy

os
argparse

imageio
PIL